Move addMusic method from the main app component to the AddMusic component to improve delegation of tasks #14
