+++ client/containers/account-settings.js
@@ -9,12 +9,12 @@ export default class AccountSettings extends Component {
   }
   handleDelete(event) {
     event.preventDefault()
-    const { navigate, deleteProfile, user: { id } } = this.props
+    const { navigate, deleteUser, user: { id } } = this.props
     const url = '/artists/' + id
     const req = { method: 'DELETE' }
     fetch(url, req)
       .then(res => res.ok)
-      .then(deleteProfile())
+      .then(deleteUser())
       .catch(err => console.error(err))
     alert('Your profile has been deleted.')
     navigate({ path: '' })
